+++ lib/cb/responses/resumes/resume_document.rb
@@ -3,17 +3,20 @@ module Cb
     class ResumeDocument < ApiResponse
       attr_accessor :resume_id, :desired_job_title, :privacy_setting, :resume_file_data, :resume_file_name, :host_site
 
-      def set_model_properties
-        @resume_id = api_response['resumeID']
-        @desired_job_title = api_response['desiredJobTitle']
-        @privacy_setting = api_response['privacySetting']
-        @resume_file_data = api_response['resumeFileData']
-        @resume_file_name = api_response['resumeFileName']
-        @host_site = api_response['hostSite']
+      def validate_api_hash
+        required_response_field(root_node, response)
       end
 
-      def required_fields
-        ['resumeID']
+      def extract_models
+        [root_node].map { |resume| ResumeDocument.new(resume) }
+      end
+
+      def hash_containing_metadata
+        response
+      end
+
+      def root_node
+        'Results'
       end
     end
   end
